Revert "refactor: home page, main container and note container"
This reverts commit d7bbc21f449a0bf66a68062377bcebb9044e5661.
