+++ src/containers/header.js
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React, { useState, useEffect, useCallback } from 'react';
 import { Form, Header, Popover } from '../components';
 import { Search } from '@styled-icons/evil/Search';
 import { PopoverContainer } from './';
