+++ src/containers/index.js
@@ -12,3 +12,4 @@ export { default as ToastContainer } from './toast';
 export { default as PaletteContainer } from './palette';
 export { default as SettingsContainer } from './settings';
 export { default as NoteContainer } from './note';
+export { MemoizedNoteContainer } from './note';
