+++ src/hooks/index.js
@@ -1,9 +1,9 @@
-export { default as useTranslateY } from './useTranslateY';
-export { default as useMatchLastSubpath } from './useMatchLastSubpath';
-export { default as useResize } from './useResize';
-export { default as useData } from './useData';
-export { default as useQuery } from './useQuery';
-export { default as useMouseClick } from './useMouseClick';
-export { default as useWindowEvent } from './useWindowEvent';
-export { default as useWindowKey } from './useWindowKey';
-export { default as useNoteHandler } from './useNoteHandler';
+export {default as useTranslateY} from "./useTranslateY";
+export {default as useMatchLastSubpath} from "./useMatchLastSubpath";
+export {default as useResize} from "./useResize";
+export {default as useData} from "./useData";
+export {default as useQuery} from "./useQuery";
+export {default as useMouseClick} from "./useMouseClick";
+export {default as useWindowEvent} from "./useWindowEvent";
+export {default as useWindowKey} from "./useWindowKey";
+export {default as useHandler} from "./useHandler";
